+++ paper/paper.bib
@@ -34,7 +34,7 @@ volume = {7},
 pages = {100978},
 year = {2020},
 issn = {2215-0161},
-doi = {https://doi.org/10.1016/j.mex.2020.100978},
+doi = {10.1016/j.mex.2020.100978},
 url = {https://www.sciencedirect.com/science/article/pii/S2215016120301989},
 author = {Dwueng-Chwuan Jhwueng},
 }
