### 2022-09-12 #4750 @kthyng
@Bisaloo Ok no problem! Can you edit your [editor page](https://joss.theoj.org/editors/120/edit) to update your availability? A comment for when you're available again would be useful, and change your availability to 0 for now.

@kavir1698 I need to waitlist your submission while we wait for an appropriate editor to become available. Thanks for your patience.
