### 2022-09-13 #4750 @mikldk
@kavir1698 - thank you for the submission.

Regarding the last sentence in Statement of need ("`EvoDynamics.jl` is being used in a couple of studies to understand the interplay between the genotypic and phenotypic architectures to understand biodiversity patterns and species coexistence."). Are those studies published or what is the status of the studies?
