+++ paper/paper.md
@@ -48,7 +48,7 @@ There is a large number of frameworks for studying genetic diversity. Currently,
 
 `EvoDynamics.jl` is a forward-time simulator based on agents with complex genotype-phenotype architectures that model complex genome processes like epistasis, pleiotropy, and gene expression in the context of stochastic mutation-recombination-migration dynamics. In addition, the agents can be connected between different types, i.e., species, producing arbitrarily complex ecological networks, and any type of species-species interactions, i.e., mutualism, commensalism, parasitism, competition. To our knowledge, this is the only framework that connects genotype-phenotype maps to biodiversity, while also accounting for complex species interactions.
 
-Being written in the Julia language, it is both high-performance and easily accessible to users to investigate the implementation of the code and to modify it, if needed. This is in contrast with most of the packages that are written in a low-level language (e.g. C++) and sometimes have an interface in an high-level language (e.g. @guillaumeNemo2006;@schiffersALADYN2014;@curratSPLATCHE32019;@bocediRangeShifter2021). Additionally, `EvoDynamics.jl` is built on top of `Agent.jl` package [@datserisAgents2022], which gives users great flexibility in defining any kind of data they want to collect. `EvoDynamics.jl` is being used in a couple of studies to understand the interplay between the genotypic and phenotypic architectures to understand biodiversity patterns and species coexistence.
+Being written in the Julia language, it is both high-performance and easily accessible to users to investigate the implementation of the code and to modify it, if needed. This is in contrast with most of the packages that are written in a low-level language (e.g. C++) and sometimes have an interface in an high-level language (e.g. @guillaumeNemo2006;@schiffersALADYN2014;@curratSPLATCHE32019;@bocediRangeShifter2021). Additionally, `EvoDynamics.jl` is built on top of `Agent.jl` package [@datserisAgents2022], which gives users great flexibility in defining any kind of data they want to collect. `EvoDynamics.jl` is being used in a couple of studies (in prep.) to understand the interplay between the genotypic and phenotypic architectures to understand biodiversity patterns and species coexistence.
 
 # Acknowledgements
 
