### 2022-09-17 5c6aaab
Update interaction rules
