+++ Project.toml
@@ -1,7 +1,7 @@
 name = "EvoDynamics"
 uuid = "c8768967-421d-4a02-8523-37736f3dbe06"
 authors = ["Ali R. Vahdati <arv.kavir@gmail.com>", "Carlos Melian"]
-version = "0.16.0"
+version = "0.17.0"
 
 [deps]
 Agents = "46ada45e-f475-11e8-01d0-f70cc89e6671"
