### 2022-09-23 #4775 @editorialbot
[REVIEW]: EvoDynamics.jl: a framework for modeling eco-evolutionary dynamics
**Submitting author:** @kavir1698 (<a href="http://orcid.org/0000-0003-0895-1495">Ali Rezaee Vahdati</a>)
**Repository:** https://github.com/kavir1698/EvoDynamics.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.17.0
