### 2022-09-24 kavir1698/EvoDynamics.jl#33 @tijeco
paramfile1.jl mutation_probabilities
It seems there is a typo in the `examples/paramfile1.jl`

https://github.com/kavir1698/EvoDynamics.jl/blob/5c6aaabe90e8a2cacd001bc047fc7bc0599aac43/examples/paramfile1.jl#L23

Needs to be changed from `mutation_probabiliti` to `mutation_probabilities`
