### 2022-09-24 #4775 @tijeco
@kavir1698 I've started going through the examples. I started with the following `agentdata, modeldata, model = runmodel("examples/paramfile1.jl")` and ran into an error, which I believe is due to a typo in the paramfile. I've raised an issue here (https://github.com/kavir1698/EvoDynamics.jl/issues/33).
