### 2022-09-24 kavir1698/EvoDynamics.jl#34 @tijeco
update output documentation
The documentation (https://kavir1698.github.io/EvoDynamics.jl/dev/) doesn't seem to have much description about what the expected output is or what the output means.

[Example 1](https://kavir1698.github.io/EvoDynamics.jl/dev/example1/)
- `modeldata` is called at the bottom of the chunk, but the actual dataframe output doesn't seem to be rendered.
[Example 2](https://kavir1698.github.io/EvoDynamics.jl/dev/example2/)
- same as example 1 in regards to `modeldata`
- also, it includes a plot, but the plot doesn't get rendered

When I run the example 2 I get the following:

```julia
(0×0 DataFrame, 31×3 DataFrame
 Row │ step   mean_fitness_per_species  species_N
     │ Int64  Vector{Float64}           Array…
─────┼──────────────────────────────────────────────
   1 │     0  [0.926653, 0.921022]      [1000, 100]
   2 │     1  [0.926646, 0.921089]      [841, 150]
  ⋮  │   ⋮               ⋮                   ⋮
  30 │    29  [0.92237, 0.929426]       [146, 255]
  31 │    30  [0.921414, 0.929547]      [76, 228]
                                     27 rows omitted, Agents.AgentBasedModel{Agents.GridSpace{2, false}, EvoDynamics.Ind{Float64, StaticArraysCore.MArray{S, Float64, 2} where S<:Tuple, StaticArraysCore.MArray{S, Bool, 2} where S<:Tuple, StaticArraysCore.MArray{S, Float64, 1} where S<:Tuple}, typeof(Agents.Schedulers.randomly), EvoDynamics.Params{Float64, Int64, String}, RandomNumbers.Xorshifts.Xoroshiro128Plus}[AgentBasedModel with 304 agents of type Ind
 space: GridSpace with size (2, 2), metric=chebyshev, periodic=false
 scheduler: randomly
 properties: ngenes, nphenotypes, growthrates, selectionCoeffs, ploidy, optvals, mutProbs, mutMagnitudes, N, E, generations, nspecies, new_N, migration_traits, vision_radius, check_fraction, migration_thresholds, step, nodes, biotic_phenotypes, abiotic_phenotypes, max_ages, names, food_sources, interactions, resources, resources_org, recombination, initial_energy, bottlenecks, repro_start, repro_end, biotic_variances, abiotic_variances, mating_schemes, phenotype_contribution_to_fitness])
```

I don't see much description on what this output means. When I search for `mean_fitness_per_species` and `species_N` in the documentation, nothing comes up.
