### 2022-09-24 #4775 @tijeco
As is, the example in `examples/paramfile2.jl` works as expected. I'd say the documentation on setting up the model parameters and running the simulations is well done, but there seems to be very little on the interpretation of what the output is.

For example, the `runmodel()` function that is used to create `agentdata, modeldata, model`,  a dataframe with values `mean_fitness_per_species` and `species_N` is generated, but I see no description of what those fields are in the documentation.

I think the documentation of the expected output needs to be updated. I've submitted an issue with more information here (https://github.com/kavir1698/EvoDynamics.jl/issues/34)
