+++ examples/paramfile1.jl
@@ -20,7 +20,7 @@ species1 = Dict(
   :expression_array => [0.28, 0.46],
   :selection_coefficient => 0.1,
   :phenotype_contribution_to_fitness => nothing,
-  :mutation_probabiliti => [0.9, 0.0, 0.0],
+  :mutation_probabilities => [0.9, 0.0, 0.0],
   :mutation_magnitudes => [0.05, 0.0, 0.0],
   :N => [100],
   :environmental_noise => 0.01,
