### 2022-09-26 kavir1698/EvoDynamics.jl#33 @kavir1698
Thank you for this. I have fixed the typo.
