+++ src/api.jl
@@ -1,4 +1,8 @@
-"Returns a tuple whose entries are the mean fitness of each species."
+"""
+    mean_fitness_per_species(model::ABM)
+
+Returns a tuple whose entries are the mean fitness of each species.
+"""
 function mean_fitness_per_species(model::ABM)
   mean_fitness = Array{Float64}(undef, model.nspecies)
   for species in 1:model.nspecies
@@ -9,6 +13,11 @@ function mean_fitness_per_species(model::ABM)
   return mean_fitness
 end
 
+"""
+    species_N(model::ABM)
+
+Returns the population size per species in an array.
+"""
 function species_N(model::ABM)
   allagents = model.agents
   if length(allagents) == 0
@@ -31,7 +40,7 @@ Creates and runs a model given `parameters`. Returns a `DataFrame` of collected 
 # Keywords
 
 * adata=[] agent data to be collected. Either agent fields or functions that accept an agent as input can be put in the array. To aggregate collected data, provide tuples inside the array. For example, to collect mean and median fitness of individuals which is in field `W`, your array will be [(:W,mean), (:W,median)].
-* mdata=[mean_fitness_per_species] model data to be collected. By default, collects mean population fitness per species. Each row of the output DataFrame corresponds to all agents and each column is the value function applied to a field. The functions in the array are applied to the model object.
+* mdata=[mean_fitness_per_species, species_N] model data to be collected. By default, collects mean population fitness per species. Each row of the output DataFrame corresponds to all agents and each column is the value function applied to a field. The functions in the array are applied to the model object.
 * when=nothing The generations from which data are collected. By default collect at all steps.
 * replicates::Int = 0 Number of replicates per simulation.
 * parallel::Bool = false Whether to run replicates in parallel. If `true`, you should add processors to your julia session (e.g. by `addprocs(n)`) and define your parameters and `EvoDynamics` on all workers. To do that, add `@everywhere` before them. For example, `@everywhere EvoDynamics`.
