+++ docs/src/example2.md
@@ -6,93 +6,153 @@ EditURL = "<unknown>/examples/example2.jl"
 
 Here we create a predator prey model of two species, one haploid and one diploid.
 
-```@example example2
+````@example example2
 using EvoDynamics
+using Plots
+````
+
+Model parameters are in a .jl file as follows:
+
+```julia
+generations = 14
+space = (2, 2)
+
+## 1. Species parameters
+
+species1 = Dict(
+  :name => "a",
+  :number_of_genes => 7,
+  :number_of_phenotypes => 4,
+  :abiotic_phenotypes => [1],
+  :biotic_phenotypes => [2, 3],
+  :migration_phenotype => 4,
+  :migration_threshold => 1.5,
+  :vision_radius => 1,
+  :check_fraction => 0.5,
+  :ploidy => 2,
+  :epistasis_matrix => [
+    1.0 0.43 -0.41 0.38 0.48 -0.43 -0.1 -0.08 -0.09 -0.5 0.41 0.44 -0.21 -0.12
+    0.34 1.0 -0.19 -0.36 0.38 -0.28 0.24 -0.22 0.12 0.12 -0.12 -0.39 0.21 0.26
+    0.05 0.27 1.0 0.04 0.01 -0.14 0.3 -0.28 0.43 -0.13 0.2 -0.02 0.25 -0.39
+    -0.12 0.33 -0.48 1.0 -0.4 -0.48 -0.22 -0.36 -0.24 -0.07 -0.12 -0.49 -0.37 0.27
+    0.25 0.25 -0.14 0.49 1.0 0.28 -0.34 -0.49 0.45 -0.14 0.26 -0.13 -0.44 -0.17
+    -0.47 0.19 -0.24 0.41 0.08 1.0 0.11 0.03 0.15 0.49 0.04 0.41 -0.19 0.13
+    0.37 0.09 -0.11 0.4 0.42 0.45 1.0 -0.01 -0.47 0.07 0.5 0.44 -0.18 -0.2
+    -0.32 0.15 0.4 -0.24 -0.21 0.5 0.22 1.0 -0.33 0.48 -0.49 0.07 0.5 -0.07
+    0.02 -0.16 0.33 0.48 -0.42 0.39 0.2 -0.11 1.0 0.46 -0.06 0.22 -0.3 0.31
+    0.41 -0.18 -0.16 -0.4 0.01 0.04 0.07 0.2 -0.37 1.0 -0.33 0.49 0.05 -0.42
+    0.03 0.25 0.14 -0.36 0.28 -0.18 0.09 -0.2 0.46 -0.48 1.0 -0.21 0.41 -0.46
+    0.0 0.44 -0.34 -0.42 0.37 -0.04 0.43 -0.25 0.21 0.19 0.29 1.0 -0.02 0.06
+    0.46 -0.1 0.14 -0.22 -0.26 0.13 -0.5 -0.41 -0.31 0.0 -0.15 0.29 1.0 0.17
+    -0.22 0.21 0.46 -0.01 -0.35 -0.11 0.25 -0.03 0.18 -0.38 -0.4 -0.28 0.05 1.0],
+  :pleiotropy_matrix => Bool[
+    1 1 0 0 0 0 0 1 0 1 0 1 0 1
+    1 0 1 1 1 0 1 1 1 0 0 1 1 0
+    1 0 1 0 1 0 0 0 0 0 1 0 1 0
+    0 0 0 0 1 0 0 1 1 1 1 0 1 0],
+  :growth_rate => 2.8,
[…]
