### 2022-09-26 kavir1698/EvoDynamics.jl#34 @kavir1698
The issue with rendering the dataframes and figures is fixed now.
I have also added a para explaining what the default output means and how to collect any other output from the simulations. The documentation strings for `mean_fitness_per_species` and `species_N` is fixed too.
