### 2022-09-30 b77f8cf
Update benchmark code
