### 2022-10-09 kavir1698/EvoDynamics.jl#35 @slwu89
Error in Wright-Fisher example
Hi @kavir1698, this issue is related to the JOSS review at [#joss](https://github.com/openjournals/joss-reviews/issues/4775).

I had a problem trying to run the 1st example at: https://kavir1698.github.io/EvoDynamics.jl/dev/example1/ Something about the parameter dict not being parsed correctly. However, the 2nd example (Predator Prey) runs with no problems and reproduces the results. I am on Julia 1.7.0, and installed the package from the Julia general repo using `]add EvoDynamics`.

```julia
julia> using EvoDynamics

julia> param_file = "/Users/slwu89/.julia/packages/EvoDynamics/j4MVr/examples/paramfile1.jl"
"/Users/slwu89/.julia/packages/EvoDynamics/j4MVr/examples/paramfile1.jl"

julia> agentdata, modeldata = runmodel(param_file);
[ Info: Loading the parameters...
ERROR: AssertionError: Species a does not have field mutation_probabilities.
Stacktrace:
 [1] check_param_names(d::Dict{Symbol, Any})
   @ EvoDynamics ~/.julia/packages/EvoDynamics/j4MVr/src/check_params.jl:11
 [2] load_parameters(paramfile::String)
   @ EvoDynamics ~/.julia/packages/EvoDynamics/j4MVr/src/load_params.jl:3
 [3] runmodel(param_file::String; adata::Nothing, mdata::Vector{Function}, when::Nothing, replicates::Int64, parallel::Bool, seeds::Nothing, agentstep::Function, modelstep::Function)
   @ EvoDynamics ~/.julia/packages/EvoDynamics/j4MVr/src/api.jl:57
 [4] runmodel(param_file::String)
   @ EvoDynamics ~/.julia/packages/EvoDynamics/j4MVr/src/api.jl:52
 [5] top-level scope
   @ REPL[3]:1
```
