### 2022-10-09 kavir1698/EvoDynamics.jl#36 @slwu89
Suggestions for JOSS paper language
Hi @kavir1698, this issue is related to the JOSS review at https://github.com/openjournals/joss-reviews/issues/4775.

I have some comments on the paper's language.

1. In the 5th paragraph in "Summary" section, I found the sentences "For example, the relationship between genetic and species diversity could not be explored accounting explicitly by the genetic-phenotypic architecture within and between each species. By accounting for the architecture within and between species, we could now explore such a connection." to be slightly hard to parse. For example, is exploring that relationship difficult because there are too many things to account for explicitly in that architecture, or a different reason? I suggest changing the wording to help make this easier to understand.
2. In the final paragraph of "Statement of Need", `Agent.jl` should be `Agents.jl`.
