### 2022-10-09 kavir1698/EvoDynamics.jl#35 @slwu89
Oh I see that the other reviewer also raised this issue! https://github.com/kavir1698/EvoDynamics.jl/issues/33 I guess the parameter files in the version hosted at the Julia general repo have not been updated yet; it works fine when cloning the repo locally. Please feel free to close this issue.
