### 2022-10-09 kavir1698/EvoDynamics.jl#37 @slwu89
Exported function documentation
Hi @kavir1698, this issue is related to the JOSS review at https://github.com/openjournals/joss-reviews/issues/4775

I see that `EvoDynamics.runmodel` is documented in the online docs. However the two other exported functions `load_parameters` and `model_initiation` are not in the online docs, I suggest adding docstrings to exported functions if users are going to interact with them.
