### 2022-10-12 kavir1698/EvoDynamics.jl#35 @kavir1698
Hi @slwu89,
Thanks for your comments. Yes, I have fixed this issue but I have not released a new version of the package yet. I will do that at the end of the review process. Meanwhile, the latest changes to the code are accessible by installing the package from the master branch: `]add EvoDynamics#master`.
