+++ paper/paper.md
@@ -37,7 +37,7 @@ Coupling genomes to phenomes is full of challenges. This is because evolutionary
 
 Evolutionary modeling have mostly focused on genes and genomic processes from a variety of angles. Ecological modeling, on the other side, have mostly focused on species level interactions with other species and the environment. `EvoDynamics.jl` attempts to connect the two by integrating genomic and phenotypic processes within and between species. Such a connection is not easy to manage analytically. To address this complexity, in-silico simulations can provide new ideas and synthesis for a wide range of researchers and interdisciplinary teams aiming to join complex processes occurring in the genotype-phenotype interface to biodiversity patterns. 
 
-`EvoDynamics.jl` can enrich the range of applications and questions previously addressed in eco-evolutionary dynamics. For example, the relationship between genetic and species diversity could not be explored accounting explicitly by the genetic-phenotypic architecture within and between each species. By accounting for the architecture within and between species, we could now explore such a connection. Similar problems like the causes of gradients in genetic and species diversity, the factors affecting population structure, and the effect of human actions on species and genetic variation [@leighOpportunities2021] can also be explored using `EvoDynamics.jl`. 
+`EvoDynamics.jl` can enrich the range of applications and questions previously addressed in eco-evolutionary dynamics. For example, the relationship between genetic and species diversity could not be explored by explicitly accounting for the genetic-phenotypic architecture within and between each species. Our package allows defining a detailed genetic-phenotypic architecture that was not possible with other packages. By accounting for the architecture within and between species, we could now explore such a connection. Similar problems like the causes of gradients in genetic and species diversity, the factors affecting population structure, and the effect of human actions on species and genetic variation [@leighOpportunities2021] can also be explored using `EvoDynamics.jl`. 
  
 
 # Statement of need
@@ -48,7 +48,7 @@ There is a large number of frameworks for studying genetic diversity. Currently,
 
 `EvoDynamics.jl` is a forward-time simulator based on agents with complex genotype-phenotype architectures that model complex genome processes like epistasis, pleiotropy, and gene expression in the context of stochastic mutation-recombination-migration dynamics. In addition, the agents can be connected between different types, i.e., species, producing arbitrarily complex ecological networks, and any type of species-species interactions, i.e., mutualism, commensalism, parasitism, competition. To our knowledge, this is the only framework that connects genotype-phenotype maps to biodiversity, while also accounting for complex species interactions.
 
-Being written in the Julia language, it is both high-performance and easily accessible to users to investigate the implementation of the code and to modify it, if needed. This is in contrast with most of the packages that are written in a low-level language (e.g. C++) and sometimes have an interface in an high-level language (e.g. @guillaumeNemo2006;@schiffersALADYN2014;@curratSPLATCHE32019;@bocediRangeShifter2021). Additionally, `EvoDynamics.jl` is built on top of `Agent.jl` package [@datserisAgents2022], which gives users great flexibility in defining any kind of data they want to collect. `EvoDynamics.jl` is being used in a couple of studies (in prep.) to understand the interplay between the genotypic and phenotypic architectures to understand biodiversity patterns and species coexistence.
[…]
