+++ src/simulation.jl
@@ -63,7 +63,8 @@ end
 """
     model_initiation(dd)
 
-Innitializes the model.
+Innitializes the model and returns an ABM object that includes all the initial agents.
+The input of the function is a dictionary returned by the `load_parameters` function.
 """
 function model_initiation(dd)
 
