### 2022-10-12 kavir1698/EvoDynamics.jl#37 @kavir1698
Hi @slwu89,
Thanks for this feedback. I have now added docstrings to the `load_parameters` and `model_initiation` functions and have included these docstrings in the online docs, on the tutorial page.
