### 2022-10-12 kavir1698/EvoDynamics.jl#36 @slwu89
I see these  comments were addressed in the commit https://github.com/kavir1698/EvoDynamics.jl/commit/c16443c5183ca7d84ec46ba14b4af108f185e44e

I'll close this issue.
