+++ paper/paper.md
@@ -29,7 +29,7 @@ bibliography: paper.bib
 
 # Summary
 
-Genotype-phenotype maps are usually complex. Studies connecting the mechanisms driving genome and phenome evolution have been done mostly in isolation. Because the rarity of studies connecting the genotype-phenotype mapping, simulations explicitly taking into account the complex genotype-phenotype architecture are incipient [@bty197]. `EvoDynamics.jl` aims to connect genomes and phenomes in an easy building-block way to allow exploring many scenarios connecting the two.
+Genotype-phenotype maps are usually complex. Studies connecting the mechanisms driving genome and phenome evolution have been done mostly in isolation. Because of the rarity of studies considering the mapping between the genotypes to phenotypes, simulations explicitly taking into account the complex genotype-phenotype architecture are incipient [@bty197]. `EvoDynamics.jl` aims to connect genomes and phenomes in an easy building-block way to allow exploring many scenarios connecting the two.
 
 Genotype-phenotype mapping occurs within a species, but between species interactions do occur in ecosystems. Yet, the genotype-to-phenotype-to-biodiversity connection is mostly unexplored. The second aim of `EvoDynamics.jl` is to tackle the genotype-phenotype coupling not only within a species but also between species. This is particularly important for understanding the role of the genetic-phenotype architecture for biodiversity response to global change. 
 
@@ -37,12 +37,12 @@ Coupling genomes to phenomes is full of challenges. This is because evolutionary
 
 Evolutionary modeling have mostly focused on genes and genomic processes from a variety of angles. Ecological modeling, on the other side, have mostly focused on species level interactions with other species and the environment. `EvoDynamics.jl` attempts to connect the two by integrating genomic and phenotypic processes within and between species. Such a connection is not easy to manage analytically. To address this complexity, in-silico simulations can provide new ideas and synthesis for a wide range of researchers and interdisciplinary teams aiming to join complex processes occurring in the genotype-phenotype interface to biodiversity patterns. 
 
-`EvoDynamics.jl` can enrich the range of applications and questions previously addressed in eco-evolutionary dynamics. For example, the relationship between genetic and species diversity could not be explored by explicitly accounting for the genetic-phenotypic architecture within and between each species. Our package allows defining a detailed genetic-phenotypic architecture that was not possible with other packages. By accounting for the architecture within and between species, we could now explore such a connection. Similar problems like the causes of gradients in genetic and species diversity, the factors affecting population structure, and the effect of human actions on species and genetic variation [@leighOpportunities2021] can also be explored using `EvoDynamics.jl`. 
+`EvoDynamics.jl` can enrich the range of applications and questions previously addressed in eco-evolutionary dynamics. For example, the relationship between genetics and species diversity could not be explored by explicitly accounting for the genetic-phenotypic architecture within and between each species. Our package allows defining a detailed genotype-phenotype architecture that was not possible with other packages. By accounting for the architecture within and between species, we can now explore such a connection. Similar problems like the causes of gradients in genetic and species diversity, the factors affecting population structure, and the effect of human actions on species and genetic variation [@leighOpportunities2021] can also be explored using `EvoDynamics.jl`. 
  
 
 # Statement of need
 
[…]
