+++ versions.js
@@ -1,5 +1,6 @@
 var DOC_VERSIONS = [
   "stable",
+  "v0.17",
   "v0.16",
   "v0.15",
   "v0.14",
@@ -12,5 +13,5 @@ var DOC_VERSIONS = [
   "v0.1",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v0.16.0";
+var DOCUMENTER_NEWEST = "v0.17.0";
 var DOCUMENTER_STABLE = "stable";
