### 2022-11-14 #4775 @Kevin-Mattheus-Moerman
@kavir1698 I am the AEiC for this track and here to help process this work for acceptance in JOSS. I have check the ZENODO archive and all seems in order there. I have read your paper too and have the below remaining points that need your attention, feel free to disagree with some of these recommendations:

- [ ] You may remove `Baltzerstrasse 6, CH-3012` (street postcode info) from the affiliation
- [ ] Consider rephrasing this sentence: `EvoDynamics.jl aims to connect genomes and phenomes in an easy building-block way to allow exploring many scenarios connecting the two.`. In particular the `in an easy building-block way` sounds awkward, and `allow exploring many scenarios connecting the two` should perhaps become something like `enabling the exploration of many scenarios connecting the two`.
- [ ] Perhaps, unless I misunderstood, `...for biodiversity response to global change...` should read `...for biodiversity responses to global change...`

Let me know when you've processed the above. Thanks.
