### 2022-11-16 171005d
Small language fixes
